requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
pydantic==2.11.5
pydantic-settings==2.9.1
python-multipart==0.0.20
dspy-ai==2.6.24
pytest==8.3.5
pytest-asyncio==0.26.0
flake8==7.2.0
black==25.1.0
python-dotenv==1.1.0
openai==1.82.0
tiktoken==0.8.0
numpy==2.2.1
rank-bm25==0.2.2
scikit-learn==1.6.0
nltk==3.9.1
click==8.1.7
PyYAML==6.0.2
